Backend/requirements.txt
distlib==0.3.4
ds-store==1.3.0
filelock==3.7.0
mac-alias==2.2.0
platformdirs==2.5.2
protobuf==3.19.4
six==1.16.0
virtualenv==20.14.1
asgiref==3.5.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
buidl==0.2.36
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.4
django-rest-knox==4.2.0
djangorestframework==3.13.1
fastecdsa==2.2.3
greenlet==1.1.2
idna==3.3
django-cors-headers==3.12.0
numpy==1.22.3
psycopg2-binary==2.9.3
pyaes==1.6.1
pycodestyle==2.8.0
pycparser==2.21
python-decouple==3.6
pytz==2022.1
requests==2.27.1
scrypt==0.8.20
SQLAlchemy==1.4.36
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9

